requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
gspread==6.1.2
google-auth==2.35.0
pydantic==2.9.2
